FIX Expand the number of items solr filters
Sets a 'large' number as the 'limit' for the Solr query; without it, the
solr filter only returns the first 20 items that match its query. Not a problem
by itself, but if solr only returns the first 20 of, say, 100 that match, and
the other filter doesn't actually match on any of those 20 (but the other 80),
the total result set ends up being 0, when it should be more.