NEW Allow OR/AND for solr facet filter inclusion
Lets the user specify whether a facet filter will 'add' more results, or prune
the number of results, with each selection.

Lets the user specify whether the counts shown for filters should show the
counts of the entire result, regardless of applied filters, or just show the
counts for the current filtered result.